Update README.md with personal statement text
whydohumanscallmecraphead

I have ADHD. Sometimes I do things that other people don't understand or think are stupid.

I wrote a piece once heralding the inception of Crapheads Anonymous. I'd like to call it CA for short but that name was already taken. I call it ChAn instead. Chan is the Chinese word meaning meditation that gives us the Japanese word Zen.

My way of living a life in which the effects of ADHD become benefits instead of holding me back is to engage in what esotericists call The Work. It is the magnum opus of the alchemists of old, the work of transformation. In The Work we transmute the impure material into alchemical gold.